kloud/domain: initial new domain api calls
